Split add_modules into 2 different regions
