+++ regions/legacy/add_modules.py
@@ -11,9 +11,6 @@ class AddModules(Region):
     _search_form_locator = (By.ID, 'collection-composer-collection-module-form-search')
     _words_field_locator = (By.CSS_SELECTOR, 'input[type="text"][name="words"]')
     _spinner_locator = (By.ID, 'kss-spinner')
-    _form_locator = (By.ID, 'collection_composer')
-    _add_workarea_checkbox_locator = (By.CSS_SELECTOR,
-                                      'input[type="checkbox"][name="selectButton"]')
 
     def __init__(self, collection, root=None):
         self.collection = collection
@@ -27,14 +24,6 @@ class AddModules(Region):
     def words_field(self):
         return self.search_form.find_element(*self._words_field_locator)
 
-    @property
-    def form(self):
-        return self.find_element(*self._form_locator)
-
-    @property
-    def add_workarea_checkbox(self):
-        return self.form.find_element(*self._add_workarea_checkbox_locator)
-
     # We need this check because this region is actually loaded via javascript
     @property
     def loaded(self):
@@ -47,37 +36,12 @@ class AddModules(Region):
         return (not self.is_element_displayed(*self._search_form_locator) and
                 not self.is_element_displayed(*self._spinner_locator))
 
-    # This is a separate `loaded` method for when performing the search (more fields show up)
-    # Necessary because without waiting for the new elements we were getting StaleElementErrors
-    @property
-    def search_loaded(self):
-        return ((self.is_element_displayed(*self._add_workarea_checkbox_locator) or
-                 'No results matched your query' in self.text) and
-                self.is_element_displayed(*self._search_form_locator) and
-                not self.is_element_displayed(*self._spinner_locator))
-
-    def wait_for_search_to_load(self):
-        """Wait for the search resuls to load."""
-        self.wait.until(lambda _: self.search_loaded)
-        return self
-
     def fill_in_search_words(self, words):
         self.words_field.send_keys(words)
         return self
 
     def submit_search(self):
         self.search_form.submit()
-        return self.wait_for_search_to_load()
-
-    def add_workarea(self):
-        self.add_workarea_checkbox.click()
-        return self
-
-    # Returns the last module added
-    def submit(self):
-        self.form.submit()
-        # Wait for the modal to close
-        self.wait_for_region_to_unload()
-        from regions.legacy.module import Module
-        module = Module(self.page, self.collection.content_nodes[-1])
-        return module.wait_for_region_to_load()
+        from regions.legacy.add_modules_with_search_results import AddModulesWithSearchResults
+        search_results = AddModulesWithSearchResults(self.collection, self.root)
+        return search_results.wait_for_region_to_load()
